+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `GettersCacheTrait::dropCachedProperty()`
+
 ### Changed
 
 - Minimum required PHP version changed from 5.4 to 5.5